+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Support for mapping Queue. A Queue will be mapped to an ArrayDeque by default. The order of elements is guaranteed to be preserved, except when the Queue is
   mapped to a Queue that inherently modifies the order of elements (e.g. PriorityQueue).
 - BeanMapper#map(Queue, Class) allowing users to map a Queue to a new Queue, mapping the elements of the source to the target class.
+- BeanMapper#map(Object[], Class) allowing users to map an array to an array with elements of the type of the target class.
 
 ### Changed